% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maildf.R
\name{convert_mail_list}
\alias{convert_mail_list}
\title{Function to convert data frame with messages to list}
\usage{
convert_mail_list(data)
}
\arguments{
\item{data}{\link[base]{data.frame} with messages}
}
\value{
\link[text.analysis]{email_list} object
}
\description{
This function is design convert a data frame with  messages to \code{email_list}
object
}
