% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{email_list}
\alias{email_list}
\title{Function to convert email data frame into a list}
\usage{
email_list(x)
}
\arguments{
\item{x}{a data frame with two columns one column for classifying emails as
ham/spam, and another column for the email messages}
}
\value{
returns an object of class \code{email_list}. This is a list with two
character vectors. One vector contains category of the email messages and
other vector contains the respective message.
}
\description{
The function \code{email_list} is used to encode a \link[base]{data.frame}
as a email_list object.
}
