% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_visuals.R
\name{explore_visuals}
\alias{explore_visuals}
\title{Visualize Text Data Exploration Results}
\usage{
explore_visuals(results)
}
\arguments{
\item{results}{A list containing the results from the explore_text_data
function.}
}
\value{
A list of ggplot objects representing the visualizations.
}
\description{
Takes the results of the explore_text_data function and creates
visualizations.
It generates plots for category distribution, message lengths, and word
counts.
}
