% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-extract_content.R
\name{extract_content}
\alias{extract_content}
\title{Creates an object to store text analyses
data.frame of n-grams words from a document}
\usage{
extract_content(input_dir, output_dir)
}
\arguments{
\item{input_dir}{A directory from which input files will be taken}

\item{output_dir}{A directory to which output files will be written}
}
\description{
Creates an object to store text analyses
data.frame of n-grams words from a document
}
