% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{final_model_df}
\alias{final_model_df}
\title{Function to create Final model matrix}
\usage{
final_model_df(data, tol)
}
\arguments{
\item{data}{cleaned corpus}

\item{tol}{a positive integer value. Used to remove the columns
which the column sum less than tolerance.}
}
\value{
final_model_mat the data in a form of a matrix can be used for the
modelling
}
\description{
The function \code{final_model_df} is designed to create the model matrix to use
modelling process. It transforms each unique word in a message into a
variable, tallying the frequency of each word. The resulting output of the
function is a data matrix where the initial column denotes the message type,
while the subsequent columns represent the words in the dataset.
}
