% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-get_filename.R
\name{get_filename}
\alias{get_filename}
\title{Extract and config a filename for saving}
\usage{
get_filename(file)
}
\arguments{
\item{file}{A file}
}
\description{
Extract and config a filename for saving
}
