% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-get_ngrams.R
\name{get_ngrams}
\alias{get_ngrams}
\title{Create data.frame of n-grams}
\usage{
get_ngrams(x, n, ...)
}
\arguments{
\item{x}{A character string of text}

\item{n}{Number of n-grams to get}

\item{...}{Additional arguments for \code{tau::textcnt}}
}
\description{
Create data.frame of n-grams
}
