% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-get_speech.R
\name{get_speech}
\alias{get_speech}
\title{Extract Parts of Speech (POS) Tagging}
\usage{
get_speech(text)
}
\arguments{
\item{text}{A \code character string of text}
}
\description{
Extract Parts of Speech (POS) Tagging
}
