% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-get_speech_text.R
\name{get_speech_text}
\alias{get_speech_text}
\title{Get objects or actions from a document}
\usage{
get_speech_text(obj, nouns = TRUE)
}
\arguments{
\item{obj}{A \code{data.frame} containing parts of speech tagging}

\item{nouns}{A \code{logical} determining if topics or actions are desired}
}
\description{
Get objects or actions from a document
}
