% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-get_text.R
\name{get_text}
\alias{get_text}
\title{Extract text from a file (.doc or .pdf)}
\usage{
get_text(input_file)
}
\arguments{
\item{input_file}{\code{character} Path to an input file}

\item{output_dir}{\code{character} Path to a directory into
which the text corpus will be saved as a list}
}
