% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_words}
\alias{get_words}
\title{Get all words from a document}
\usage{
get_words(doc)
}
\arguments{
\item{doc}{A character string extracted from a document}
}
\description{
Get all words from a document
}
