% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_accuracy.R
\name{log_classification}
\alias{log_classification}
\title{Logistic Regression Classification Accuracy Measures}
\usage{
log_classification(data, threshold = 0.5)
}
\arguments{
\item{data}{list of length 2 from \link[text.analysis]{partition}}

\item{threshold}{probability for classification}
}
\value{
\link[base]{list} with length 3
}
\description{
This function is designed to compute accuracy measures for email data set
using Logistic Regression classification model. This function return the
model, confusion matrix and accuracy measures.
}
