% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messages.R
\name{lower_case}
\alias{lower_case}
\title{Function to make all messages to lower case}
\usage{
lower_case(x)
}
\arguments{
\item{x}{a \link[text.analysis]{email_list} object}
}
\value{
a \link[text.analysis]{email_list} object with lower case
characters.
}
\description{
The function \code{lower_case} is design to make character vectors of
\link[text.analysis]{email_list} from upper to lower case to lower case
character vectors.
}
