% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_accuracy.R
\name{nb_classification}
\alias{nb_classification}
\title{Naive - Bayes Classification Accuracy Measures}
\usage{
nb_classification(data, laplace = 0, ...)
}
\arguments{
\item{data}{list of length 2 from \link[text.analysis]{partition}}

\item{laplace}{description positive double controlling Laplace smoothing.
The default (0) disables Laplace smoothing.}

\item{...}{futher parameters from \link[e1071]{naiveBayes} function}
}
\value{
\link[base]{list} with length 3
}
\description{
This function is designed to compute accuracy measures for email data set
using Naive Bayes classification model. This function return the model,
confusion matrix and accuracy measures.
}
