% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_partitioning.R
\name{partition}
\alias{partition}
\title{Data partitioning for train and test cross validation}
\usage{
partition(model_df, prob)
}
\arguments{
\item{model_df}{final data frame created using \code{final_model_df}
function}

\item{prob}{probability of ham/spam messages to include in the train data.}
}
\value{
a list of both train and test data sets.
}
\description{
The partition function is designed to create training and testing data
matrices using the model matrix generated from final_model_matrix. These
matrices are then used to train and validate the fitted model.
}
