% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maildf.R
\name{read_maildf}
\alias{read_maildf}
\title{Function to Read Email Message Data File}
\usage{
read_maildf(file, header = FALSE, sep = "", skip = 0, quote = "\\"", ...)
}
\arguments{
\item{file}{the file name from which the data is to be read. Each row of
the table is represented as one line in the file. If the file name does not
include an absolute path, it is assumed to be relative to the current working
directory obtained using \link[base]{getwd}.}

\item{header}{a boolean value indicating whether the file includes variable
names as its first line. If not provided, it is inferred from the file
format: the \code{header} is set to \code{TRUE} only if the first row has one fewer
field than the total number of columns.}

\item{sep}{this parameter used to separate fields in the file. Each line's
values are split using this character. \code{sep = ""}, is the default separator
and it defaults to 'white space', meaning one or more spaces, tabs, newlines,
or carriage returns.}

\item{skip}{an integer indicating the number of lines in the data file to
skip before starting to read the data.}

\item{quote}{specifies the quoting characters used in the file. To disable
quoting entirely, set \code{quote = ""}. Refer to scan for how embedded quotes are
handled. Quoting applies only to columns read as character, unless
\code{colClasses} is specified.}

\item{...}{Further arguments to be passed to \link[utils]{read.table}.}
}
\value{
\link[base]{data.frame}
}
\description{
This function is design to read data file that contains email messages.
Primarily this function parse a table-formatted file and convert it into a
list with two character vectors, where one vector corresponds to a category
of the mail and the other vector corresponds to the respective emails.
}
