% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messages.R
\name{remove_punctuations}
\alias{remove_punctuations}
\title{Function to make remove all punctuations from messages}
\usage{
remove_punctuations(x)
}
\arguments{
\item{x}{a \link[text.analysis]{email_list} object}
}
\value{
a \link[text.analysis]{email_list} object with only letters
characters.
}
\description{
The function \code{remove_punctuations} is design to remove all punctuations from
\link[text.analysis]{email_list} character vectors.
}
