% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messages.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Function to remove stop words from messages}
\usage{
remove_stopwords(x)
}
\arguments{
\item{x}{a \link[text.analysis]{email_list} object}
}
\value{
a \link[text.analysis]{email_list} object without stop words
characters.
}
\description{
The function \code{remove_stopwords} is design to remove all  stop words
from \link[text.analysis]{email_list} character vectors.
}
