% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sms_spam_collection}
\alias{sms_spam_collection}
\title{SMS Spam Collection}
\format{
\subsection{\verb{SMS Spam Collection}}{

A data frame with 5,574 rows and 2 columns:
\describe{
\item{category}{Spam or Ham}
\item{message}{Actual Message}
}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/228/sms+spam+collection}
}
\usage{
sms_spam_collection
}
\description{
The SMS Spam Collection is a public set of SMS labeled messages that have
been collected for mobile phone spam research.
}
\keyword{datasets}
