% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMSSpamCollection}
\alias{SMSSpamCollection}
\title{SMS Spam Collection}
\format{
\subsection{\code{SMSSpamCollection}}{

A data frame with 5,574 rows and 2 columns:
\describe{
\item{V1}{Spam or Ham}
\item{V2}{Actual Message}
}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/228/sms+spam+collection}
}
\usage{
SMSSpamCollection
}
\description{
The SMS Spam Collection is a public set of SMS labeled messages that have been collected for mobile phone spam research.
}
\keyword{datasets}
