% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_accuracy.R
\name{svm_classification}
\alias{svm_classification}
\title{SVM Classification Accuracy Measures}
\usage{
svm_classification(
  data,
  kernel = "radial",
  gamma = 1/ncol(data$train),
  cost = 1,
  ...
)
}
\arguments{
\item{data}{list of length 2 from \link[text.analysis]{partition}}

\item{kernel}{the kernel used in training and predicting. You might consider
changing some of the following parameters, depending on the kernel type.}

\item{gamma}{parameter needed for all kernels except linear
(default: 1/(data dimension))}

\item{cost}{cost of constraints violation (default: 1)—it is the
'C'- constant of the regularization term in the Lagrange formulation.}

\item{...}{futher parameters from \link[e1071]{svm} function}
}
\value{
\link[base]{list} with length 3
}
\description{
This function is designed to compute accuracy measures for email data set
using Support Vector Machine classification model. This function return the
model, confusion matrix and accuracy measures.
}
