% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maildf.R
\name{convert_mail_list}
\alias{convert_mail_list}
\title{Function to convert data frame with messages to list}
\usage{
convert_mail_list(data)
}
\arguments{
\item{data}{\link[base]{data.frame} with messages}
}
\value{
\link[text.analysis]{email_list} object
}
\description{
This function is design convert a data frame with  messages to \code{email_list}
object
}
\examples{
# raw data as a data frame
df_raw <- data.frame(
  type = c("ham", "spam", "ham", "ham", "spam"),
  message = c(
    "Hey there! how is your day going",
    "You have won a free vacation click here to claim your prize 2",
    "Reminder meeting today do not forget",
    "Just wanted to say hi and see how you are doing",
    "Your account has been compromised click here to update your password"
  )
)

# convert the data frame into a list
convert_mail_list(df_raw)

}
