% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_data}
\alias{explore_data}
\title{Explore Text Data for email_list}
\usage{
explore_data(email_list_obj)
}
\arguments{
\item{email_list_obj}{An object of class \code{email_list}.}
}
\value{
A list containing the results of the data exploration.
}
\description{
This function performs initial exploratory data analysis on text data
contained within an email_list object.
It returns a list containing summary statistics, distributions of categories,
message lengths, word counts, and checks for missing values.
}
