% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_visuals.R
\name{explore_visuals}
\alias{explore_visuals}
\title{Visualize Text Data Exploration Results}
\usage{
explore_visuals(data, numwords = 10)
}
\arguments{
\item{data}{a cleaned \link[text.analysis]{email_list} object}

\item{numwords}{number of words need to plot for each category (default: 10)}
}
\value{
pie chart for category and barcharts for each category
}
\description{
This function use to create visualizations category distribution and
number of words for each category.
}
