% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_visuals.R
\name{explore_visuals}
\alias{explore_visuals}
\title{Visualize Text Data Exploration Results}
\usage{
explore_visuals(data, numwords = 10)
}
\arguments{
\item{data}{a cleaned \link[text.analysis]{email_list} object}

\item{numwords}{number of words need to plot for each category (default: 10)}
}
\value{
pie chart for category and barcharts for each category
}
\description{
This function use to create visualizations category distribution and
number of words for each category.
}
\examples{
# raw data as a data frame
df_raw <- data.frame(
  type = c("ham", "spam", "ham", "ham", "spam"),
  message = c(
    "Hey there! how is your day going",
    "You have won a free vacation click here to claim your prize 2",
    "Reminder meeting today do not forget",
    "Just wanted to say hi and see how you are doing",
    "Your account has been compromised click here to update your password"
  )
)

df_raw

# input data list using `convert_mail_list`
ls <- convert_mail_list(df_raw)

# clean raw data to get the corpus
ls_lower <- lower_case(ls)
ls_nonumber <- remove_numbers(ls_lower)
ls_nopunc <- remove_punctuations(ls_nonumber)
ls_nospace <- remove_whitespaces(ls_nopunc)
corpus <- remove_stopwords(ls_nospace)

# Visualization
explore_visuals(corpus)

}
