% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_accuracy.R
\name{log_classification}
\alias{log_classification}
\title{Logistic Regression Classification Accuracy Measures}
\usage{
log_classification(data, threshold = 0.5)
}
\arguments{
\item{data}{list of length 2 from \link[text.analysis]{partition}}

\item{threshold}{probability for classification}
}
\value{
\link[base]{list} with length 3
}
\description{
This function is designed to compute accuracy measures for email data set
using Logistic Regression classification model. This function return the
model, confusion matrix and accuracy measures.
}
\examples{
\dontrun{
# input data list "df"
df <- list(
  type = c("ham", "spam", "ham", "ham", "spam"),
  message = c(
    "hey there how is your day going",
    "you have won a free vacation click here to claim your prize",
    "reminder meeting today do not forget",
    "just wanted to say hi and see how you are doing",
    "your account has been compromised click here to update your password"
  )
)

# create train and test data using partition function
model_df <- final_model_df(df, 2)
train_test_ls <- partition(model_df, 0.6)

# run Logistic Regression Model
log_classification(train_test_ls)
}

}
