% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messages.R
\name{lower_case}
\alias{lower_case}
\title{Function to make all messages to lower case}
\usage{
lower_case(x)
}
\arguments{
\item{x}{a \link[text.analysis]{email_list} object}
}
\value{
a \link[text.analysis]{email_list} object with lower case
characters.
}
\description{
The function \code{lower_case} is design to make character vectors of
\link[text.analysis]{email_list} from upper to lower case to lower case
character vectors.
}
\examples{
# raw data as a data frame
df_raw <- data.frame(
  type = c("ham", "spam", "ham", "ham", "spam"),
  message = c(
    "Hey there! how is your day going",
    "You have won a free vacation click here to claim your prize 2",
    "Reminder meeting today do not forget",
    "Just wanted to say hi and see how you are doing",
    "Your account has been compromised click here to update your password"
  )
)

# input data list using `convert_mail_list`
ls <- convert_mail_list(df_raw)

# convert all the letters to lower case
lower_case(ls)

}
