% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messages.R
\name{remove_numbers}
\alias{remove_numbers}
\title{Function to make remove all numbers from messages}
\usage{
remove_numbers(x)
}
\arguments{
\item{x}{a \link[text.analysis]{email_list} object}
}
\value{
a \link[text.analysis]{email_list} object with only letters
characters.
}
\description{
The function \code{remove_numbers} is design to remove all numbers from
\link[text.analysis]{email_list} character vectors.
}
