% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messages.R
\name{remove_punctuations}
\alias{remove_punctuations}
\title{Function to remove all punctuations from messages}
\usage{
remove_punctuations(x)
}
\arguments{
\item{x}{a \link[text.analysis]{email_list} object}
}
\value{
a \link[text.analysis]{email_list} object with only letters
characters.
}
\description{
The function \code{remove_punctuations} is design to remove all punctuations from
\link[text.analysis]{email_list} character vectors.
}
\examples{
# raw data as a data frame
df_raw <- data.frame(
  type = c("ham", "spam", "ham", "ham", "spam"),
  message = c(
    "Hey there! how is your day going",
    "You have won a free vacation click here to claim your prize 2",
    "Reminder meeting today do not forget",
    "Just wanted to say hi and see how you are doing",
    "Your account has been compromised click here to update your password"
  )
)

# input data list using `convert_mail_list`
ls <- convert_mail_list(df_raw)

# remove all the punctuations
remove_punctuations(ls)

}
