% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_messages.R
\name{remove_whitespaces}
\alias{remove_whitespaces}
\title{Function to make remove extra white spaces from messages}
\usage{
remove_whitespaces(x)
}
\arguments{
\item{x}{a \link[text.analysis]{email_list} object}
}
\value{
a \link[text.analysis]{email_list} object with only letters
characters.
}
\description{
The function \code{remove_whitespaces} is design to remove all extra white spaces
from \link[text.analysis]{email_list} character vectors.
}
