% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_accuracy.R
\name{rf_classification}
\alias{rf_classification}
\title{Random Forest Classification Accuracy Measures}
\usage{
rf_classification(data, ntree = 500, mtry = sqrt(ncol(data$train)), ...)
}
\arguments{
\item{data}{list of length 2 from \link[text.analysis]{partition}}

\item{ntree}{Number of trees to grow. This should not be set to too small a
number, to ensure that every input row gets predicted at least a few times.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
Note that the default values are different for classification (sqrt(p)
where p is number of variables in x)}

\item{...}{futher parameters from \link[randomForest]{randomForest} function}
}
\value{
\link[base]{list} with length 3
}
\description{
This function is designed to compute accuracy measures for email data set
using Random Forest classification model. This function return the model,
confusion matrix and accuracy measures.
}
\examples{
\dontrun{
# input data list "df"
df <- list(
  type = c("ham", "spam", "ham", "ham", "spam"),
  message = c(
    "hey there how is your day going",
    "you have won a free vacation click here to claim your prize",
    "reminder meeting today do not forget",
    "just wanted to say hi and see how you are doing",
    "your account has been compromised click here to update your password"
  )
)

# create train and test data using partition function
model_df <- final_model_df(df, 2)
train_test_ls <- partition(model_df, 0.6)

# run RandomForest Classifier
rf_classification(train_test_ls)
}

}
