% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordcloud.R
\name{split_spamHam}
\alias{split_spamHam}
\title{Function to split into spam and ham subsets}
\usage{
split_spamHam(data)
}
\arguments{
\item{data}{corpus file}
}
\description{
This function splits the corpus into two subsets: spam and ham.
}
