% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordcloud.R
\name{wordcloud_all}
\alias{wordcloud_all}
\title{Function to Create Word Clouds for all dataset}
\usage{
wordcloud_all(data)
}
\arguments{
\item{file}{clean text data file}
}
\description{
This function will explore the dataset with word clouds. It will provide a
visualization of the frequency of words in our corpus.
}
\details{
Exploring the dataset with word clouds. This will show the frequency of words
in our dataset. The larger the font size is in the word cloud, the more
frequently a word appears.
}
