% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordcloud.R
\name{wordcloud_ham}
\alias{wordcloud_ham}
\title{Function to Create Word Clouds for all Ham messages}
\usage{
wordcloud_ham(ham_set)
}
\arguments{
\item{ham_set}{clean ham data file}
}
\description{
This function will explore the "ham" dataset with word clouds. It will
provide a visualization of the frequency of words in our corpus.
}
\details{
Exploring the dataset with word clouds. This will show the frequency of words
in our "ham" dataset. The larger the font size is in the word cloud, the more
frequently a word appears.
}
